Split the API idempotency contract into business-key and idempotency-key
AbstractIdempotencyContract was named for idempotency but every test drove withBusinessKey, and its own Javadoc admitted it. Rename it to AbstractBusinessKeyContract to match what it pins, paralleling the store tier's AbstractActiveBusinessKeyContract.

The real withIdempotencyKey permanent-merge behavior had no portable contract; only the RI-specific IdempotencyKeyIT covered it. Add a new AbstractIdempotencyContract that mandates merge as the spec: a duplicate idempotency key must return the original job's handle, must not start a second execution, and the reservation must survive completion and ignore a changed task. Wire up the RI subclass RiIdempotencyKeyIT and both catalog entries, and drop the now-redundant standalone IdempotencyKeyIT. Verified green on the wildfly-managed plus postgresql cell.
